Added elb health check when using ASG and moving instance out of standby
This does the same check of waiting for the instance to be InService in the ELB
as when we're working without ASGs. In my tests I had deployments that failed
the ELB health check but the deployment was wrongly marked as successful as the
ASG exit standby function doesn't seem to check the ELB health check.

I added the function autoscaling_group_elbs in commons_functions to get all the
ELBs attached to a ASG.